move blockchain functionality into blockchain function
